Refactor functionality into separate files
- merge
- preprocessing
- yaml_node
