+++ pkg/utils/meta/diff.go
@@ -10,8 +10,13 @@ import (
 
 	"github.com/elliotchance/orderedmap/v3"
 	"go.yaml.in/yaml/v4"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 )
+
+// section represents a single section in a manifest file (either a manifest or a comment)
+type section struct {
+	key     string
+	content []byte
+}
 
 func newSection(key string, content []byte) *section {
 	return &section{
@@ -20,13 +25,13 @@ func newSection(key string, content []byte) *section {
 	}
 }
 
-type section struct {
-	key     string
-	content []byte
-}
-
 func (s *section) isComment() bool {
 	return len(s.key) > 0 && s.key == string(s.content)
+}
+
+// manifestDiff holds the three versions of manifest maps used in three-way merge
+type manifestDiff struct {
+	oldDefault, newDefault, current *orderedmap.OrderedMap[string, []byte]
 }
 
 // ThreeWayMergeManifest creates or updates a manifest based on a given YAML object.
@@ -81,288 +86,26 @@ func ThreeWayMergeManifest(oldDefaultYaml, newDefaultYaml, currentYaml []byte) (
 	return postProcess(output), nil
 }
 
-func threeWayMergeSection(oldDefaultYaml, newDefaultYaml, currentYaml []byte) ([]byte, error) {
-	// Parse all three versions
-	var oldDefault, newDefault, current yaml.Node
-	if err := yaml.Unmarshal(newDefaultYaml, &newDefault); err != nil {
-		return nil, err
+func newManifestDiff(oldDefaultYaml, newDefaultYaml, currentYaml []byte) (*manifestDiff, error) {
+	md := &manifestDiff{}
+	var err error
+	if md.oldDefault, err = splitManifestFile(oldDefaultYaml); err != nil {
+		return nil, fmt.Errorf("parsing oldDefault file for manifest diff failed: %w", err)
 	}
-	if err := yaml.Unmarshal(currentYaml, &current); err != nil {
-		return nil, err
+	if md.newDefault, err = splitManifestFile(newDefaultYaml); err != nil {
+		return nil, fmt.Errorf("parsing newDefault file for manifest diff failed: %w", err)
 	}
-
-	// If no old default exists, use empty node (will cause all existing keys to be treated as user-added)
-	if len(oldDefaultYaml) > 0 {
-		if err := yaml.Unmarshal(oldDefaultYaml, &oldDefault); err != nil {
-			return nil, err
-		}
+	if md.current, err = splitManifestFile(currentYaml); err != nil {
+		return nil, fmt.Errorf("parsing current file for manifest diff failed: %w", err)
 	}
 
-	return encodeResult(threeWayMerge(&oldDefault, &newDefault, &current))
+	// Handle single manifest files with different names/namespaces
+	md.normalizeSingleManifestKeys()
+
+	return md, nil
 }
 
-func encodeResult(merged *yaml.Node) ([]byte, error) {
-	var buf bytes.Buffer
-	encoder := yaml.NewEncoder(&buf)
-	defer encoder.Close()
-	encoder.SetIndent(2)
-	encoder.CompactSeqIndent()
-	if err := encoder.Encode(merged); err != nil {
-		return nil, err
-	}
-
-	return buf.Bytes(), nil
-}
-
-// threeWayMerge performs a three-way merge of YAML nodes
-// oldDefault: the previous default template
-// newDefault: the new default template
-// current: the user's current version (possibly modified)
-func threeWayMerge(oldDefault, newDefault, current *yaml.Node) *yaml.Node {
-	// Unwrap document nodes
-	if oldDefault.Kind == yaml.DocumentNode {
-		oldDefault = oldDefault.Content[0]
-	}
-	if newDefault.Kind == yaml.DocumentNode {
-		newDefault = newDefault.Content[0]
-	}
-	if current.Kind == yaml.DocumentNode {
-		return &yaml.Node{
-			Kind:    yaml.DocumentNode,
-			Content: []*yaml.Node{threeWayMerge(oldDefault, newDefault, current.Content[0])},
-		}
-	}
-
-	// If current equals oldDefault (including comments), no user modifications were made - use newDefault
-	if nodesEqual(oldDefault, current, true) {
-		return newDefault
-	}
-
-	// Build maps for easier lookup (we only handle mappings for Kubernetes manifests)
-	oldMap := buildMap(oldDefault)
-	currentMap := buildMap(current)
-	newMap := buildMap(newDefault)
-
-	// Create result node preserving current's comments and style
-	result := &yaml.Node{
-		Kind:        yaml.MappingNode,
-		Style:       current.Style,
-		Tag:         newDefault.Tag,
-		HeadComment: current.HeadComment,
-		LineComment: current.LineComment,
-		FootComment: current.FootComment,
-	}
-
-	// Process keys from newDefault
-	for i := 0; i < len(newDefault.Content); i += 2 {
-		newKeyNode, newValueNode := newDefault.Content[i], newDefault.Content[i+1]
-		key := newKeyNode.Value
-		oldValue, oldExists := oldMap[key]
-		currentValue, currentExists := currentMap[key]
-
-		var resultKeyNode, resultValue *yaml.Node
-
-		if oldExists && !currentExists {
-			// Has been dropped from current.
-			continue
-		}
-		if !currentExists {
-			// New key - add from newDefault
-			resultKeyNode, resultValue = newKeyNode, newValueNode
-		} else {
-			resultKeyNode = findKeyNode(current, key)
-
-			// Handle nested structures (mappings and sequences)
-			switch {
-			case currentValue.Kind == yaml.MappingNode && newValueNode.Kind == yaml.MappingNode:
-				if !oldExists {
-					oldValue = &yaml.Node{Kind: yaml.MappingNode}
-				}
-				resultValue = threeWayMerge(oldValue, newValueNode, currentValue)
-			case currentValue.Kind == yaml.SequenceNode && newValueNode.Kind == yaml.SequenceNode:
-				if !oldExists {
-					oldValue = &yaml.Node{Kind: yaml.SequenceNode}
-				}
-				resultValue = threeWayMergeSequence(oldValue, newValueNode, currentValue)
-			case oldExists && !nodesEqual(oldValue, newValueNode, false):
-				resultValue = &yaml.Node{
-					Kind: newValueNode.Kind, Value: newValueNode.Value, Style: newValueNode.Style, Tag: newValueNode.Tag,
-					HeadComment: currentValue.HeadComment, LineComment: currentValue.LineComment, FootComment: currentValue.FootComment,
-					Content: newValueNode.Content,
-				}
-			default:
-				resultValue = currentValue
-			}
-		}
-
-		result.Content = append(result.Content, resultKeyNode, resultValue)
-	}
-
-	// Then add any keys from current that don't exist in newDefault AND didn't exist in oldDefault (user-added keys)
-	for i := 0; i < len(current.Content); i += 2 {
-		keyNode, valueNode := current.Content[i], current.Content[i+1]
-		key := keyNode.Value
-
-		_, existsInNew := newMap[key]
-		_, existedInOld := oldMap[key]
-
-		if !existsInNew && !existedInOld {
-			// key exists only in current (user-added) - keep it at the end
-			result.Content = append(result.Content, keyNode, valueNode)
-		}
-	}
-
-	return result
-}
-
-// threeWayMergeSequence performs a three-way merge of YAML sequence nodes (arrays)
-// Order is preserved based on newDefault, with user additions appended at the end
-func threeWayMergeSequence(oldDefault, newDefault, current *yaml.Node) *yaml.Node {
-	if nodesEqual(oldDefault, current, true) {
-		return newDefault
-	}
-
-	result := &yaml.Node{
-		Kind:        yaml.SequenceNode,
-		Style:       newDefault.Style,
-		Tag:         newDefault.Tag,
-		HeadComment: current.HeadComment,
-		LineComment: current.LineComment,
-		FootComment: current.FootComment,
-	}
-
-	// Build sets for lookup
-	oldSet := make(map[string]bool)
-	for _, item := range oldDefault.Content {
-		oldSet[nodeToString(item)] = true
-	}
-
-	currentMap := make(map[string]bool)
-	for _, item := range current.Content {
-		currentMap[nodeToString(item)] = true
-	}
-
-	newSet := make(map[string]bool)
-	for _, item := range newDefault.Content {
-		newSet[nodeToString(item)] = true
-	}
-
-	// Process items in current order first to preserve order.
-	for _, currentItem := range current.Content {
-		key := nodeToString(currentItem)
-		if !oldSet[key] || newSet[key] {
-			// Add item if it has not been removed in newDefault
-			result.Content = append(result.Content, currentItem)
-		}
-	}
-
-	// Add new items from newDefault that don't exist in current or old.
-	for _, newItem := range newDefault.Content {
-		key := nodeToString(newItem)
-		if !oldSet[key] && !currentMap[key] {
-			// New template item - add from newDefault
-			result.Content = append(result.Content, newItem)
-		}
-	}
-
-	return result
-}
-
-// nodeToString converts a node to a string representation for comparison
-func nodeToString(node *yaml.Node) string {
-	if node.Kind == yaml.ScalarNode {
-		return node.Value
-	}
-	// For non-scalar nodes, marshal to YAML for comparison
-	var buf bytes.Buffer
-	encoder := yaml.NewEncoder(&buf)
-	encoder.SetIndent(2)
-	utilruntime.Must(encoder.Encode(node))
-	utilruntime.Must(encoder.Close())
-	return buf.String()
-}
-
-// findKeyNode finds the key node for a given key in a mapping (assumes node is a mapping)
-func findKeyNode(node *yaml.Node, key string) *yaml.Node {
-	for i := 0; i < len(node.Content); i += 2 {
-		if node.Content[i].Value == key {
-			return node.Content[i]
-		}
-	}
-	return &yaml.Node{Kind: yaml.ScalarNode, Value: key}
-}
-
-// buildMap creates a map from YAML mapping node for easier lookup (assumes node is a mapping)
-func buildMap(node *yaml.Node) map[string]*yaml.Node {
-	result := make(map[string]*yaml.Node)
-	for i := 0; i < len(node.Content); i += 2 {
-		key := node.Content[i].Value
-		value := node.Content[i+1]
-		result[key] = value
-	}
-	return result
-}
-
-// nodesEqual checks if two YAML nodes are equal
-// compareComments: if true, comments must also match; if false, only values are compared
-func nodesEqual(a, b *yaml.Node, compareComments bool) bool {
-	if a.Kind != b.Kind {
-		return false
-	}
-
-	// Check comments if requested
-	if compareComments {
-		if a.HeadComment != b.HeadComment ||
-			a.LineComment != b.LineComment ||
-			a.FootComment != b.FootComment {
-			return false
-		}
-	}
-
-	switch a.Kind {
-	case yaml.ScalarNode:
-		return a.Value == b.Value
-	case yaml.SequenceNode:
-		if len(a.Content) != len(b.Content) {
-			return false
-		}
-		// For sequences, always compare in order
-		for i := range a.Content {
-			if !nodesEqual(a.Content[i], b.Content[i], compareComments) {
-				return false
-			}
-		}
-		return true
-	case yaml.MappingNode:
-		if len(a.Content) != len(b.Content) {
-			return false
-		}
-		if compareComments {
-			// When comparing comments, order matters
-			for i := range a.Content {
-				if !nodesEqual(a.Content[i], b.Content[i], true) {
-					return false
-				}
-			}
-		} else {
-			// When ignoring comments, use map comparison (order-independent)
-			aMap := buildMap(a)
-			bMap := buildMap(b)
-			if len(aMap) != len(bMap) {
-				return false
-			}
-			for key, aValue := range aMap {
-				bValue, exists := bMap[key]
-				if !exists || !nodesEqual(aValue, bValue, false) {
-					return false
-				}
-			}
-		}
-		return true
-	}
-	return true
-}
-
+// splitManifestFile splits a multi-document YAML file into separate manifests
 func splitManifestFile(combinedYaml []byte) (*orderedmap.OrderedMap[string, []byte], error) {
 	var values [][]byte
 	if len(combinedYaml) > 0 { // Only split if there is content
@@ -383,27 +126,36 @@ func splitManifestFile(combinedYaml []byte) (*orderedmap.OrderedMap[string, []by
 	return om, nil
 }
 
-type manifestDiff struct {
-	oldDefault, newDefault, current *orderedmap.OrderedMap[string, []byte]
+// buildKey builds a unique key for a manifest using apiVersion/kind/namespace/name
+func buildKey(t map[string]any) string {
+	typeKey := buildTypeKey(t)
+	metadata, _ := t["metadata"].(map[string]any)
+	name, _ := metadata["name"].(string)
+	namespace, _ := metadata["namespace"].(string)
+	if typeKey == "" && namespace == "" && name == "" {
+		return ""
+	}
+
+	return typeKey + "/" + namespace + "/" + name
 }
 
-func newManifestDiff(oldDefaultYaml, newDefaultYaml, currentYaml []byte) (*manifestDiff, error) {
-	md := &manifestDiff{}
-	var err error
-	if md.oldDefault, err = splitManifestFile(oldDefaultYaml); err != nil {
-		return nil, fmt.Errorf("parsing oldDefault file for manifest diff failed: %w", err)
+// buildTypeKey builds a type key from apiVersion and kind
+func buildTypeKey(t map[string]any) string {
+	apiVersion, _ := t["apiVersion"].(string)
+	kind, _ := t["kind"].(string)
+	if apiVersion == "" || kind == "" {
+		return ""
 	}
-	if md.newDefault, err = splitManifestFile(newDefaultYaml); err != nil {
-		return nil, fmt.Errorf("parsing newDefault file for manifest diff failed: %w", err)
-	}
-	if md.current, err = splitManifestFile(currentYaml); err != nil {
-		return nil, fmt.Errorf("parsing current file for manifest diff failed: %w", err)
-	}
+	return apiVersion + "/" + kind
+}
 
-	// Handle single manifest files with different names/namespaces
-	md.normalizeSingleManifestKeys()
-
-	return md, nil
+// extractTypeKey extracts "apiVersion/kind" from a manifest YAML
+func extractTypeKey(yamlContent []byte) string {
+	var t map[string]any
+	if err := yaml.Unmarshal(yamlContent, &t); err != nil {
+		return ""
+	}
+	return buildTypeKey(t)
 }
 
 // normalizeSingleManifestKeys handles the case where all three files contain a single manifest
@@ -495,81 +247,4 @@ func collectAppendix(diff *manifestDiff) []*section {
 		}
 	}
 	return appendix
-}
-
-// extractTypeKey extracts "apiVersion/kind" from a manifest YAML
-func extractTypeKey(yamlContent []byte) string {
-	var t map[string]any
-	if err := yaml.Unmarshal(yamlContent, &t); err != nil {
-		return ""
-	}
-	return buildTypeKey(t)
-}
-
-func buildTypeKey(t map[string]any) string {
-	apiVersion, _ := t["apiVersion"].(string)
-	kind, _ := t["kind"].(string)
-	if apiVersion == "" || kind == "" {
-		return ""
-	}
-	return apiVersion + "/" + kind
-}
-
-func buildKey(t map[string]any) string {
-	typeKey := buildTypeKey(t)
-	metadata, _ := t["metadata"].(map[string]any)
-	name, _ := metadata["name"].(string)
-	namespace, _ := metadata["namespace"].(string)
-	if typeKey == "" && namespace == "" && name == "" {
-		return ""
-	}
-
-	return typeKey + "/" + namespace + "/" + name
-}
-
-// keepLeftAlignedMarker is a marker to identify left-aligned comment lines during pre- and post-processing.
-const keepLeftAlignedMarker = "###KEEP_LEFT_ALIGNED###"
-
-type lineProcessor func([]byte, *bytes.Buffer)
-
-func process(yamlContent []byte, processLine lineProcessor) []byte {
-	if len(yamlContent) == 0 {
-		return yamlContent
-	}
-	buf := bytes.Buffer{}
-	lines := bytes.Split(yamlContent, []byte("\n"))
-	for i, line := range lines {
-		processLine(line, &buf)
-		if i < len(lines)-1 {
-			buf.Write([]byte("\n"))
-		}
-	}
-	return buf.Bytes()
-}
-
-// preProcessLine adds a marker to a left aligned comment line.
-// As the "go.yaml.in/yaml/v4" package does not store the original indentation of comments in the node model,
-// they are indented during marshaling. This marker helps to identify such lines for left-alignment during post-processing.
-func preProcessLine(line []byte, buf *bytes.Buffer) {
-	buf.Write(line)
-	if bytes.HasPrefix(line, []byte("#")) {
-		buf.Write([]byte(keepLeftAlignedMarker))
-	}
-}
-
-// postProcessLine removes the marker added during pre-processing and left-aligns the comment line again after marshaling.
-func postProcessLine(line []byte, buf *bytes.Buffer) {
-	if before, ok := bytes.CutSuffix(line, []byte(keepLeftAlignedMarker)); ok {
-		line = before
-		line = bytes.TrimLeft(line, " ")
-	}
-	buf.Write(line)
-}
-
-func preProcess(yamlContent []byte) []byte {
-	return process(yamlContent, preProcessLine)
-}
-
-func postProcess(yamlContent []byte) []byte {
-	return process(yamlContent, postProcessLine)
 }
