+++ pkg/utils/meta/merge.go
@@ -1,0 +1,184 @@
+// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package meta
+
+import (
+	"go.yaml.in/yaml/v4"
+)
+
+// threeWayMergeSection performs a three-way merge on a single YAML section
+func threeWayMergeSection(oldDefaultYaml, newDefaultYaml, currentYaml []byte) ([]byte, error) {
+	// Parse all three versions
+	var oldDefault, newDefault, current yaml.Node
+	if err := yaml.Unmarshal(newDefaultYaml, &newDefault); err != nil {
+		return nil, err
+	}
+	if err := yaml.Unmarshal(currentYaml, &current); err != nil {
+		return nil, err
+	}
+
+	// If no old default exists, use empty node (will cause all existing keys to be treated as user-added)
+	if len(oldDefaultYaml) > 0 {
+		if err := yaml.Unmarshal(oldDefaultYaml, &oldDefault); err != nil {
+			return nil, err
+		}
+	}
+
+	return encodeResult(threeWayMerge(&oldDefault, &newDefault, &current))
+}
+
+// threeWayMerge performs a three-way merge of YAML nodes
+// oldDefault: the previous default template
+// newDefault: the new default template
+// current: the user's current version (possibly modified)
+func threeWayMerge(oldDefault, newDefault, current *yaml.Node) *yaml.Node {
+	// Unwrap document nodes
+	if oldDefault.Kind == yaml.DocumentNode {
+		oldDefault = oldDefault.Content[0]
+	}
+	if newDefault.Kind == yaml.DocumentNode {
+		newDefault = newDefault.Content[0]
+	}
+	if current.Kind == yaml.DocumentNode {
+		return &yaml.Node{
+			Kind:    yaml.DocumentNode,
+			Content: []*yaml.Node{threeWayMerge(oldDefault, newDefault, current.Content[0])},
+		}
+	}
+
+	// If current equals oldDefault (including comments), no user modifications were made - use newDefault
+	if nodesEqual(oldDefault, current, true) {
+		return newDefault
+	}
+
+	// Build maps for easier lookup (we only handle mappings for Kubernetes manifests)
+	oldMap := buildMap(oldDefault)
+	currentMap := buildMap(current)
+	newMap := buildMap(newDefault)
+
+	// Create result node preserving current's comments and style
+	result := &yaml.Node{
+		Kind:        yaml.MappingNode,
+		Style:       current.Style,
+		Tag:         newDefault.Tag,
+		HeadComment: current.HeadComment,
+		LineComment: current.LineComment,
+		FootComment: current.FootComment,
+	}
+
+	// Process keys from newDefault
+	for i := 0; i < len(newDefault.Content); i += 2 {
+		newKeyNode, newValueNode := newDefault.Content[i], newDefault.Content[i+1]
+		key := newKeyNode.Value
+		oldValue, oldExists := oldMap[key]
+		currentValue, currentExists := currentMap[key]
+
+		var resultKeyNode, resultValue *yaml.Node
+
+		if oldExists && !currentExists {
+			// Has been dropped from current.
+			continue
+		}
+		if !currentExists {
+			// New key - add from newDefault
+			resultKeyNode, resultValue = newKeyNode, newValueNode
+		} else {
+			resultKeyNode = findKeyNode(current, key)
+
+			// Handle nested structures (mappings and sequences)
+			switch {
+			case currentValue.Kind == yaml.MappingNode && newValueNode.Kind == yaml.MappingNode:
+				if !oldExists {
+					oldValue = &yaml.Node{Kind: yaml.MappingNode}
+				}
+				resultValue = threeWayMerge(oldValue, newValueNode, currentValue)
+			case currentValue.Kind == yaml.SequenceNode && newValueNode.Kind == yaml.SequenceNode:
+				if !oldExists {
+					oldValue = &yaml.Node{Kind: yaml.SequenceNode}
+				}
+				resultValue = threeWayMergeSequence(oldValue, newValueNode, currentValue)
+			case oldExists && !nodesEqual(oldValue, newValueNode, false):
+				resultValue = &yaml.Node{
+					Kind: newValueNode.Kind, Value: newValueNode.Value, Style: newValueNode.Style, Tag: newValueNode.Tag,
+					HeadComment: currentValue.HeadComment, LineComment: currentValue.LineComment, FootComment: currentValue.FootComment,
+					Content: newValueNode.Content,
+				}
+			default:
+				resultValue = currentValue
+			}
+		}
+
+		result.Content = append(result.Content, resultKeyNode, resultValue)
+	}
+
+	// Then add any keys from current that don't exist in newDefault AND didn't exist in oldDefault (user-added keys)
+	for i := 0; i < len(current.Content); i += 2 {
+		keyNode, valueNode := current.Content[i], current.Content[i+1]
+		key := keyNode.Value
+
+		_, existsInNew := newMap[key]
+		_, existedInOld := oldMap[key]
+
+		if !existsInNew && !existedInOld {
+			// key exists only in current (user-added) - keep it at the end
+			result.Content = append(result.Content, keyNode, valueNode)
+		}
+	}
+
+	return result
+}
+
+// threeWayMergeSequence performs a three-way merge of YAML sequence nodes (arrays)
+// Order is preserved based on newDefault, with user additions appended at the end
+func threeWayMergeSequence(oldDefault, newDefault, current *yaml.Node) *yaml.Node {
+	if nodesEqual(oldDefault, current, true) {
+		return newDefault
+	}
+
+	result := &yaml.Node{
+		Kind:        yaml.SequenceNode,
+		Style:       newDefault.Style,
+		Tag:         newDefault.Tag,
+		HeadComment: current.HeadComment,
+		LineComment: current.LineComment,
+		FootComment: current.FootComment,
+	}
+
+	// Build sets for lookup
+	oldSet := make(map[string]bool)
+	for _, item := range oldDefault.Content {
+		oldSet[nodeToString(item)] = true
+	}
+
+	currentMap := make(map[string]bool)
+	for _, item := range current.Content {
+		currentMap[nodeToString(item)] = true
+	}
+
+	newSet := make(map[string]bool)
+	for _, item := range newDefault.Content {
+		newSet[nodeToString(item)] = true
+	}
+
+	// Process items in current order first to preserve order.
+	for _, currentItem := range current.Content {
+		key := nodeToString(currentItem)
+		if !oldSet[key] || newSet[key] {
+			// Add item if it has not been removed in newDefault
+			result.Content = append(result.Content, currentItem)
+		}
+	}
+
+	// Add new items from newDefault that don't exist in current or old.
+	for _, newItem := range newDefault.Content {
+		key := nodeToString(newItem)
+		if !oldSet[key] && !currentMap[key] {
+			// New template item - add from newDefault
+			result.Content = append(result.Content, newItem)
+		}
+	}
+
+	return result
+}
